requirements.txt
asgiref==3.8.1
Django==4.2.13
djangorestframework==3.15.1
factory-boy==3.3.0
Faker==25.5.0
install==1.3.5
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.1
